frontend: refactor to recommended redux structure
Refactor the app to use the feature-based structure redux recommends.
