+++ frontend/src/app/Main.tsx
@@ -5,12 +5,12 @@ import { Drawer } from "react-native-drawer-layout";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import FloatingButton from "./components/FloatingButton";
-import Map from "./components/Map";
-import Sheet from "./components/Sheet";
-import RouteList from "./screens/RouteList";
-import LayoutStyle from "./style/layout";
-import SpacingStyle from "./style/spacing";
+import FloatingButton from "../common/components/FloatingButton";
+import LayoutStyle from "../common/style/layout";
+import SpacingStyle from "../common/style/spacing";
+import Map from "../features/map/Map";
+import Sheet from "../features/navi/Sheet";
+import RouteList from "../features/routes/RouteList";
 
 /**
  * Controls the percentage of the screen taken up by the bottom sheet in
